+++ src/irvine/oeis/a078/A078725.java
@@ -14,6 +14,6 @@ public class A078725 extends LambdaSequence {
 
   /** Construct the sequence. */
   public A078725() {
-    super(0, n -> Puma.primeZ(Z.valueOf(n * n + n * 1).divide2().add(Z.FOUR)));
+    super(0, n -> Puma.primeZ(Z.valueOf((long) n * n + n * 1L).divide2().add(Z.FOUR)));
   }
 }
