+++ src/irvine/oeis/a113/A113257.java
@@ -14,6 +14,6 @@ public class A113257 extends LambdaSequence {
 
   /** Construct the sequence. */
   public A113257() {
-    super(1, n -> Integers.SINGLETON.sum(1, n, i -> Z.valueOf(i * i).pow(Z.valueOf(n - i + 1).square())));
+    super(1, n -> Integers.SINGLETON.sum(1, n, i -> Z.valueOf((long) i * i).pow(Z.valueOf(n - i + 1).square())));
   }
 }
