+++ src/irvine/oeis/a117/A117266.java
@@ -18,6 +18,6 @@ public class A117266 extends LambdaSequence {
   public A117266() {
     super(0, n -> (n <= 1) ? Z.valueOf(n + 1) : ((n == 2) ? Z.SIX : Rationals.SINGLETON.sum(0, n, k ->
       new Q(MemoryFactorial.SINGLETON.factorial(2 * (n - k)),
-        MemoryFactorial.SINGLETON.factorial(n - k).square()).multiply(Z.ONE.shiftLeft((n - k) * (n + k - 3) / 2))).num()));
+        MemoryFactorial.SINGLETON.factorial(n - k).square()).multiply(Z.ONE.shiftLeft((long) (n - k) * (n + k - 3) / 2))).num()));
   }
 }
