+++ src/irvine/oeis/a121/A121676.java
@@ -15,6 +15,6 @@ public class A121676 extends LambdaSequence {
 
   /** Construct the sequence. */
   public A121676() {
-    super(0, n -> (n == 0) ? Z.ONE : Integers.SINGLETON.sum(0, n + 1, k -> Binomial.binomial(n + 1, k).multiply(Binomial.binomial((n - 1) * k, n - k))));
+    super(0, n -> (n == 0) ? Z.ONE : Integers.SINGLETON.sum(0, n + 1, k -> Binomial.binomial(n + 1, k).multiply(Binomial.binomial((long) (n - 1) * k, n - k))));
   }
 }
