+++ src/irvine/oeis/a121/A121680.java
@@ -13,6 +13,6 @@ public class A121680 extends LambdaSequence {
 
   /** Construct the sequence. */
   public A121680() {
-    super(0, n -> Integers.SINGLETON.sum(0, n + 1, k -> Binomial.binomial(n + 1, k).multiply(Binomial.binomial((n + 1) * k, n - k))));
+    super(0, n -> Integers.SINGLETON.sum(0, n + 1, k -> Binomial.binomial(n + 1, k).multiply(Binomial.binomial((long) (n + 1) * k, n - k))));
   }
 }
